requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0

# Configuration Management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Progress & Logging
tqdm>=4.65.0

# Experiment Tracking (optional)
wandb>=0.15.0

# Data Version Control (optional)
dvc[gdrive]>=3.0.0

# Testing
pytest>=7.4.0

# API & UI (for later phases)
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.25.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
